create new header subfolder for Content and refactor props to types file
