+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContentHeader/index.ts
@@ -1,0 +1,1 @@
+export { ComboboxContentHeader } from "./ComboboxContentHeader";
